[cov] add dbg_print_unittest_variant
This commit adds a cc_library `dbg_print_unittest_variant`, which is a
variant of the `dbg_print` library built with `DBG_PRINT_UNIT_TEST_`
defined.

The existing `dbg_print_unittest` is modified to depend on
`dbg_print_unittest_variant` rather than directly including the source files.

This change enables coverage to be collected for the `dbg_print` module in the
unit tests, since test sources are by default uninstrumented.

Change-Id: Ifd9cdc950a6ca92bf91e84b538426436d6ce20e1
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>