[dv,otp_ctrl] Fix some unit test failures
Allow scoreboard to wait some cycles for an expected timeout to become an
outgoing alert to account for flops on the path.

The otp_macro fsm errors are flagged as bus integrity errors, which sounds
wrong and perhaps should be fixed to be flagged as fatal_prim_otp_alert.
For now I changed the order of checks in task check_sec_cm_fi_resp, in
otp_ctrl_common_vseq.sv.tpl.

Templetize function sec_cm_fi_ctrl_svas since the buffered partitions depend
on per-top config. The function was also incorrect for earlgrey.

Signed-off-by: Guillermo Maturana <matute@zerorisc.com>