[rom_ctrl,rtl] Drop address of returned data from rom_ctrl_counter
The reason to look at this is that I want to parameterise the response
latency of the ROM itself. Since the counter doesn't contain the ROM,
it feels wrong for it to know or care about the ROM's responses.

Working through this carefully has made me realise that I didn't
design it very cleanly in the first place! The slightly rejigged
version here should also be a bit simpler to reason about.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>