[cryptotest] Switch security level to low for KMAC functest
lowRISC/opentitan#27673 introduced device security checks that are
enabled when the key.securitry_level > kOtcryptoKeySecurityLevelLow.

As this check fails on FPGAs (the corresponding ROM does not enable
some of the security features), set
key.security_level = kOtcryptoKeySecurityLevelLow for the
kmac_functest. This does not influence the KMAC testing as currently
the security level is not used within the KMAC driver..

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>