[rom_ctrl,rtl] Teach rom_ctrl_mux about 2-cycle version of ROM
The mux passes rom_rvalid_i back to the bus (in bus_rvalid_o), so
needs to know the ROM latency because it needs to know exactly when
the selection switches between the FSM and the bus.

Conveniently, we have both registered versions of sel_bus_i and have
checked whether each is Mubi4True anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>