[otbn] Patch an error in the OTBN boot services program
There is an error in the OTBN boot services program loaded by the ROM
that can affect the random share values.  This bug is caused by a
simple transposition of operands in the `p256_base` source code.

Since the OTBN program is stored in ROM and loaded into OTBN by the ROM,
we patch it in the immutable section before we use that portion of
the program.

DO NOT cherrypick this change to the `master` branch.

Signed-off-by: Chris Frantz <cfrantz@google.com>