[ci] Retry CI Curl downloads on failure
We sometimes see transient networking failures causing mid-stream errors
that cannot be easily handled when downloading e.g. Verilator or
Verible. These are often caused by issues connecting with Github
servers.

This commit wraps these `curl` commands with a mechanism that retries,
with exponential backoff, to handle some of these transient failures.
This means that after the first failure, the script will wait 2s and
retry, then 4s, then 8s and finally 16s. If the 5th attempt still fails,
then the download will fail.

A preferable long-term mechanism would be to mirror our dependencies to
a more reliable download source, but this change is intended to make CI
more robust to transient failures in the meantime (and we could still
get transient networking errors even with a mirror).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>