[crypto,sw] Add the affine is on curve check to ECDSA base point mult
This commit adds the existing affine is on curve check to
base point multiplications in ECDSA. This is done as a FI
countermeasure.
This commit also adds a wrapper function that does the comparison
between the right hand and the left hand side of the result of
the is on curve check. This way we save some IMEM.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>