[rom_ext] Do not clear `RLB` in the ROM_EXT
In the ES chip, we cleared Rule Locking Bypass (RLB) to make locked
entries permanent because we needed to lock out access to a couple of
peripheral regions.

In the A1/A2 chips, we want to give the owner full control over the
ePMP.  It is the owner's responsibility to configure the MSECCFG
register, which includes the RLB, MMWP and MML configuration bits.

Signed-off-by: Chris Frantz <cfrantz@google.com>