[top,otp_ctrl_mmap] Remove declared partition sizes
Sizes of partitions are determined by some attributes plus item(s)
it contains. If a partition has a declared size it takes precedence,
so a mismatch can cause either an error because it doesn't fit its
items, or memory that is not labeled, and potentially unused.

Remove declared partition sizes for all partitions with items.

Signed-off-by: Guillermo Maturana <matute@zerorisc.com>