[otp_ctrl, dv] Unrecognized hierarchical name  in $asseroff task
Xcelium complains about the $assertoff assertion control task having
OtpErrorState_A. The $assertoff task have a hierarchy that contains
a path to a generate loop as well. But assertion control on generate
blocks is not supported by default and the switch added in this PR
enables the support for them.

Signed-off-by: Kinza Qamar Zaman <kqzaman@gmail.com>