[sw][usb_logging] Remove unneeded `ctype.h` include
As part of an effort to remove unneeded toolchain dependencies, I
noticed that `sw/device/lib/testing/usb_logging.c` was including
`ctype.h` even though we don't link against the C standard library.
This patch renames the custom `isprintable` function to match the
standard `isprint` name and signature. For now, we'll keep it in this
file. It can be moved to a centralized place later, if we find the need
for it elsewhere.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>
(cherry picked from commit c0bd25995ef2d6a6b5f1a6996ab96602a136c5b6)