[ottf,console] Make backends optional
Previously, the console code was assuming that every possible backend
(UART,SPI) is available. This implicitely requires that the top has
both a UART and a SPI device. With this commit, the console code
now only compiles the backend for which the IP is present. Since
it is now possible to have no backend, the code will create a null
backend if no other is available.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>