[bazel] use `rctx.getenv` instead of repository `environ`
Bazel documentation says that `environ` is deprecated. This change
removes the usage of it and replace `rctx.os.environ` with `rctx.getenv`
which is the recommended way that automatically registered a dependency.

This change also fixes the missing dependency on `HOME` and `PATH`.

Suggested-by: Amaury Pouly <amaury.pouly@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>
(cherry picked from commit ea788d6ec2bfd58085356fcfbba938c245deac7f)