[hw,bitstream] Mark two targets as test-only
These targets directly depend on bitstream targets. If those targets
(which are created by the bitstream cache) are cached entries then this
is fine because those are not marked as test-only. However if there is
no cache entry for them, they will point to the fusesoc target for
the bitstream which is marked as test-only. This creates a non-test-only
to test-only dependency which is an error.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>