[crypto, rsa] Fix public-key exponent to F4
F4=2^16+1 is the de-facto standard public-key exponent for all
standardized RSA encryption/signature schemes. The cryptolib supported
a haphazard implementation for arbitrary exponents that reverted back
to F4 in many places. This commit removes all the residuals with
regards to arbitrary values in preparation for the upcoming RSA
hardening efforts.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>