[dv,dma] Improve SoC System bus testing
Support finer granularity in the choice of the 4GiB address
window into which the TL-UL memory is mapped. Instead of being
4GiB-aligned it is now 4B-aligned. This just provides more
robust checking of the addressing arithmetic within the DUT.

Introduce separate 4GiB windows with different base addresses
for the source and destination traffic.

Since the DMA controller itself can only perform transfers
of less than 4GiB, now that the adapter supports separated
read and writes transfers we can lose the 'waiving' of full
verification.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>