[usb,tests] Print log at crucial points during the suspend test
Presently, the only points where the harness displays the device
log is when synchronizing the phase. If on-device logging is enabled,
this can produce very confusing traces where a lot of USB hosts messages
are followed by lots of USB device messages which occures *much* earlier!
This makes the interpretation of the log difficult.

This commit introduces a mechanism to let the harness display the device
log for a short duration at certain crucial points during the tests. The
delay is configurable and can be set to 0 if one does not want to introduce
delays during the test.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit d97a5ae369067b7dcd788f80cf1ca9e77561179e)