[test] Disable CSRNG alert during KAT test
The KAT test generates the all-zero genbits twice in a row (before and
after reseeding). CSRNG has a recoverable alert that detects when the
same genbits were generated twice as it indicates an attack.

We have to disable this alert for the second generation and restore it
afterwards.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>
(cherry picked from commit 364b6e88843e405dab526dafc030a586c75d8cf0)