[entropy_src/dv] Correct computation of window_size/rng_frames in SCB
To check the size of the noise samples queue, we need to know the
expected window size, which depends on whether we're running in FIPS
or bypass mode, and whether the single-bit mode is enabled. The reason
is that as of commit a0f1e73ac7093fbb5a972bd973e62d78a1d43a60, the
window size in bypass mode is specified in bits. In contrast, the window
size in FIPS mode is always specified in symbols and switching to
single-bit mode effectively alters the symbol size.

This commit helps resolving around 85% of the failures currently
observed in ENTROPY_SRC block-level DV.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>