[dv,dma] Check that CFG_REGWEN protects configuration
Attempt to change the configuration registers that are controlled
by `CFG_REGWEN` whilst the DUT is operating. This hardware-
controlled write enable cannot be stimulated by the CIP register
testing, so we attempt to modify them whilst the DUT is performing
a long transfer.

Using longer transfers merely gives us more time to test the
significant number of registers protected by this write enable.

Extend the scoreboard to include modeling of the CFG_REGWEN
functionality; ignore writes when disabled by the DUT.
Note that presently we must read from the `CFG_REGWEN` register
of the DUT to update the mirrored value since the DUT controls
the state of this write enable automatically.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>