[reggen, systemrdl] Change from PeakRDL exporter to lowRISC exporter
Swap the PeakRDL exporter for a lowRISC exporter due to limitations on
the first. Then we no longer need to depend on the whole PeakRDL. Only the
SystemRDL compiler suffices.
The PeakRDL exporter does not support parameters, therefore we would not
be able to export the local parameters from the hjson to RDL.

Signed-off-by: Douglas Reis <doreis@lowrisc.org>