[sw] Remove `compile_commands.json` generation script
This script has been broken for a while because of a problem with `bazel
aquery --config=riscv32` picking up Python files which are not supported
on the device side.

The databases that it generates are far too large (tens of GiBs) to be
usable given `clangd` loads them into memory.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>