[crypto] Remove decoys in hardened_* functions
In the previous implementation of random_order, the length returned
was larger than the length of the buffer we want to traverse. Hence,
decoys were introduced and they were used for the additional
iterations.

lowRISC/opentitan#27875 modified the random_order function. Now,
the length returned is identical to the length of the buffer we
want to traverse. Hence, no additional iterations are executed -
so we can remove the decoys.

As described in lowRISC/opentitan#27638 decoys are anyways not
particularly as a SCA countermeasure.

Closes lowRISC/opentitan#27638

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 1136ecb45e11d9c355fc3ca05969ddc89d54224a)