[sw,ujson] Make ujson_ottf a dual cc library
This library is depended upon by several json libraries. Unfortunately
those libraries are used both by the device code, and by host code for
bindgen. This creates a host dependency on ujson_ottf which now brings
in the entire ottf_console. To avoid this issue, make it a dual cc
library and break the dependency for host. This is not really a
satisfactory solution and in the future, this library should probably
be reworked.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>