[entropy_src/dv] Fix health test threshold computation functions
Before this commit, the threshold computation functions for the adaptive
proportion, the bucket and the Markov health tests was broken. As a
result, the important entropy_src_rng and entropy_src_rng_max_rate tests
using this function could not exercise the second half of the
ENTROPY_SRC hardware pipeline (from the conditioner onward) as the
final processing of the conditioner is only triggered if a window passes
the health checks. Because of the wrong thresholds, practically all
health tests constantly failed for these three health tests.

This commit fixes the computation of the thresholds and improves the
documentation of the corresponding functions.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>