[sw,cryptolib] Adapt internal mult and add new test
This commit changes the internal multiplicaiton such that
the src and dst regs of the sel instruction are not the same.
Furthermore, this commit adds a new test for the internal
multiplication, since the scalar_mult_test will be using
the blinded version of the internal multiplication and will
thus no longer test the basic internal multiplication which
is still needed for the sign algorithm.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>