[sw,e2e] Add ROM testcases for debug info OTP switches
This change adds ROM e2e tests for the debug info OTP switches.

The `OWNER_SW_CFG_ROM_RSTMGR_INFO_EN` OTP item controls two features:
- Alert info collection.
- CPU info collection.

This change adds two testcases to verify the behaviors for these two switches:
- `alert_info_test.c`: Test whether the alert info can be collected after an
  alert escalation.
- `cpu_info_test.c`: Test whether the CPU info can be collected after an
  unexpected memory access.

The tests ensure that when the corresponding OTP bit is enabled, the info
can be collected, and when disabled, the info cannot be collected.

---

Errata #27498: On Earlgrey, a typo in ROM sets an incorrect register when the
CPU debug info OTP switch is enabled. The `exit_success` is temporarily set to
the failure string for the Earlgrey branch to keep the test happy.

Change-Id: I951943ddf01c595b1da7131fba06eee184ac1aaa
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>