[racl_ctrl,dv] Start predicting ERROR_LOG / ERROR_LOG_ADDRESS
This isn't all that complicated. The only problem is that we mustn't
update a register prediction at the same time as the register is being
accessed. This uses a couple of scoreboard prediction
values (error_log_pred and error_log_address_pred) for checking
register read values.

To avoid things "drifting" between the scoreboard and the RAL, we also
have a process for each register that tries to keep the RAL prediction
in sync.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>