[sw,test] create QEMU I2C test
This is a test for the EG 1.0.0 I2C implementation.

This test is QEMU-specific, as QEMU does not support pinmuxing which is
used in all hardware tests.

There is a limited selection of applicable I2C devices in QEMU to test.
For this test, the DS1338 RTC+NVRAM was chosen for testing host
functionality, and the i2c-echo device was chosen for testing host and
target functionality. The test expects these devices to be present at
the same addresses on each of the three I2C buses.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>