[opentitantool] Updating opentitantool to add private_sign_key
This additional options added to the opentitantool cmd line allow for
skipping signing of image if the option is set to false. There is a case
for this option since there can be a signed.bin file and we dont want to
re-sign.

Signed-off-by: esnguyen <sarzanguyen@google.com>
(cherry picked from commit b58dae581a55bda6192fa3c84af235f41c0d2a05)