[bazel] Add support for unpopulated manifest extension slots
This change enables the manifest rules to accept empty string `""` for
unpopulated extension slots.

Since the input type is a string list, which doesn't allow `None` values,
empty strings are used to represent unpopulated extension slots.

If no extensions are explicitly provided, a default set will be used, as was
done previously.

Change-Id: I23bc12818b77d0d5632e45fc283231714d8d20f6
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>