[sw,e2e] Add an e2e test for a bad SPX manifest extension
This commit adds an end-to-end test case for a bad SPX manifest.

Specifically, it places the `spx_signature` extension in an incorrect slot
within the manifest extension list. The test expects the ROM_EXT to
correctly detect this error and return `kErrorManifestBadExtension`.

This test case improves coverage for the scenario where an `spx_key` is present,
but `spx_signature` is missing.

Change-Id: If0207f0efbbd216148253097b802811d971c6f4f
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>