[entropy_src/dv] Correctly model saturation behavior in healt tests
Especially for the Darjeeling variant and when large window sizes are
used, it is possible that the prim_sum_tree counters in the adaptive
proportion and the Markov tests overflow. The primitive got fixed to
saturate instead in the previous commit. This commit aligns the
ENTROPY_SRC scoreboard accordingly.

This is related to lowRISC/OpenTitan#27471.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>