[opentitanproxy] remove mio usage completely and replace with async
This keeps most logic intact, further refactoring left to subsequent commits.
The only major change made here is that the actual handling of `Transport` makes
use of a mpsc channel and handled in a single async block. This is used instead
of a `tokio::sync::Mutex` solution as `Transport` cannot be `Send`.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>