[test,keymgr] Always initialize keymgr in keymgr_derive_cdi test
This test was changed so that it only called
`keymgr_testutils_initialize` on the first reset (i.e. power up), to
avoid the flash being reinitialized on the second boot (which caused an
alert to trigger, which was then caught).

However, in non-SiVal environments (boot stage != owner), the initialize
function calls `keymgr_testutils_init_nvm_then_reset` to do this logic,
which already checks the reset reason to determine if flash should be
programmed, so this conditional is redundant.

Importantly, the `keymgr_testutils_intialize` function also performs a
lot of important keymgr intialization which is required for the test to
pass, causing the test to now fail in the non-Sival environments.

This was not an issue for SiVal environments which would not write the
flash and thus would fully initialize the keymgr on power up.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>