[rom_ctrl,dv] Rewrite a rom_ctrl vseq to avoid tl_access_sub
This task was really intended to be internal, so maybe it's better not
to use it directly. Upsettingly, avoiding that means you have to
duplicate a bit of its contents, but that turns out to be cleaner than
I feared.

I've also expanded out the error messages and (hopefully) made it a
bit more obvious what the sequence is doing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>