[sw,e2e] Add fault test for OWNER_SW returning to ROM_EXT.
The `OWNER_SW` is not intended to return control to the `ROM_EXT`; if it does,
the `ROM_EXT` should report an error.

This commit introduces a new fault test, `return_boot_failed`, which simulates
the `OWNER_SW` returning to `ROM_EXT` (e.g., via a `ret` instruction) to verify
that `ROM_EXT` correctly handles this unexpected scenario.

This test can make sure the `ROM_EXT` handles this error situation correctly.

Change-Id: I741557b00cea523bb46b96c375b95740e0e65372
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>