[sw,keymgr] trap if keymgr is idle after starting op
The `keymgr_wait_until_done()` function should be only called after
an operation has been started by the caller.

This commit removes the `IDLE` status check from the switch-case in
`keymgr_wait_until_done()` to enforce this invariant. If the keymgr
is found to be `IDLE` the execution will now proceed to the default
error-handling case and trigger a trap.

Change-Id: I35b5107174796e9a52eff7ab33bdc2260453a73b
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>