[sw,test] Add tests for unlocking from recovery state
This change adds two tests: `unlock_when_recovery_state_test` and
`unlock_with_owner_recovery_key_test`.

The first test checks that when the device is in the recovery state, an
unlock operation can proceed by using the owner recovery key. This ensures
that even in a recovery scenario, the device can be unlocked for subsequent
ownership transfers.

The second test ensures that an unlock operation using the owner recovery
key works as expected, regardless of the initial state, provided the device
can transition to an unlocked state.

Change-Id: Ib15f2319ef81363e24942d347a51b09fc2297927
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>