[opentitanproxy] use a tokio task for receiving messages
This prepares for the upcoming change to handle async messages directly without
having to continous poll the transport.

This also gets rid of the redundant `ProxyLookupError` -- not only it's an
extra unneeded step, it is also incorrect if the DNS resolves to both IPv6 and Ipv4
and only one of them works.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>