[prim,alert] Fix SkewCycles assertions in alert sender & receiver
The assertions weren't quite right after the change in f5cdd1ce and
the SkewCycles parameter ended up being used to measure the time until
the signal integrity problem was reported, rather than the time the
signals had to be equal.

This fixes things again. Since I'm touching the code in question, it
also switches the assertions to use a more conventional style (using
default clocking and default disable iff to avoid the reasonably
inpenetrable `ASSERT macro) and slightly reformats a couple of the
multi-cycle properties to make them a bit easier to read.

It turns out that the `ASSERT macros expand to nothing when the tool
doesn't support assertions (on Verilator, for example), so I had to
shuffle some "ifdef INC_ASSERT" lines around a little to get the
equivalent effect. But I think it's still clearer.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>