[otp_ctrl/rtl] Unconditionally reject DAI writes to zeroization marker
Before this commit, DAI writes to the zeroization marker field were
rejected for buffered partitions with a HW digest and for unbuffered
partitions.  While this effectively applies to all OTP partitions
configured upstream (for both Darjeeling and Earlgrey) already, it is
cleaner to just *always* reject DAI writes to the zeroization marker
field.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>