[fpga/splice] Correct MMI order for updatemem
...and undo the sequence of a transformation followed by its inverse.

Updatemem fills the memories based on the order of the bit lanes as they
appear in the MMI file. In a VMEM file, the MSBs appear first, but in
the MMI file, the LSBs appear first. gen_vivado_mem_image corrects for
an ostensibly broken MMI generation script. Correct the MMI script and
remove the transformations in gen_vivado_mem_image.

Signed-off-by: Alex Williams <awill@opentitan.org>