[sw/build] Switch failing AES tests to cw340
This commit switches tests that are failing due to AES being used
in the unmasked configuration to cw340 where masking still is
enabled.
edn_sw_mode doesn't make sense to run with an unused mask configuration
since this test stalls AES by not providing entropy. However if AES
is not in the masked configuration, entropy is not needed and
AES cannot be stalled in this way.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>