[bazel,cov] Add transition to disable coverage collection
This change allows users to disable coverage collection for particular Bazel
targets. This can be useful for targets like metal ROM, which are sensitive
to changes in code size.

Support for ROM/ROM_EXT coverage will be added in later PRs;
to ensure the minimal framework works, ROM/ROM_EXT coverage collection is
explicitly disabled in this commit.

Change-Id: Icbfa8f64fa8a580d6f595046035a25731d5dc6af
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>