[bazel,cov] Add coverage tool and elf to runfiles
The coverage collection tool in Bazel currently cannot access its own runfiles;
it can only access the test's runfiles.

To make those tools and the ELF file available for generating the coverage report,
This change adds the coverage tool (the `cc_toolchain`) and the ELF binary into the
runfiles of the `opentitan_binary` and `opentitan_test` rules.

Change-Id: I72acb0b43e3ac53dee260a0517ae8a38a6dcab20
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>