[sw,cov] Save coverage report in opentitantool UartConsole
This change adds a feature to opentitantool's `UartConsole` that allows saving
device-side coverage reports.

The new implementation parses coverage data embedded in the console output,
validates it using CRC, and then saves it to a `.xprofraw` file.

This is part of Bazel coverage integration to collect reports for tests running
directly on the OpenTitan device. It follows LLVM's convention for saving to the
path specified by the `LLVM_PROFILE_FILE` environment variable, which Bazel will
set.

Additionally, a new `wait_for_coverage` function is introduced, allowing
`opentitantool` to block until coverage data is received or skipped.

Change-Id: I682655aeeb3c16228036ad58fb41187f7eb932cb
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>