[opentitanlib] Handle unexpected starts in UART RX monitoring
In the UART bitbanging RX decoding logic, where a GPIO monitoring
interface is used for the UART RX signal, it may be the case that we
started monitoring the pin either when (a) RX had not yet been pulled up
by the other UART, so the line starts floating / low or (b)
mid-transaction, meaning that we will start reading garbage, and the
decoder may stay out of sync if the UART does not return idle for some
time.

To avoid these issues, introduce additional sampling/decoding logic to
handle both these cases. Check that the first edge is a falling edge,
and skip to the next edge if not. Before waiting for this falling edge,
first wait for some "stable state" where we can take a consistent number
of samples longer than a UART frame. These can all be logical low (break
condition) or high (idle), as we just care that the UART is not
mid-transmission when we start decoding.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit 737f1629bd6737a5035e80858867006b0f691f8f)